attempt regression tests
