+++ .github/workflows/regression.yml
@@ -1,0 +1,90 @@
+name: Regression
+
+on:
+  push:
+    branches: [ "master" ]
+  pull_request:
+    branches: [ "master" ]
+
+jobs:
+  Main:
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: [ "3.11" ]
+
+    steps:
+    - name: Checkout
+      uses: actions/checkout@v3
+
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install pytest
+        pip install .
+    - name: Test with pytest
+      run: |
+        python -m pytest test/test_root.py test/test_process.py
+  
+  CTR:
+    runs-on: ubuntu-latest
+      container:
+        image: docker://devkitpro/devkitarm
+        options: --cpus 2
+
+      steps:
+        - name: Checkout
+          uses: actions/checkout@v3
+
+        - name: Install dependencies
+          run: |
+            python -m pip install --upgrade pip
+            python -m pip install pytest
+            pip install .
+
+        - name: Test with pytest
+          run: |
+            python -m pytest test/test_process_ctr.py
+
+  HAC:
+    runs-on: ubuntu-latest
+
+      container:
+        image: docker://devkitpro/devkita64
+        options: --cpus 2
+      
+      steps:
+        - name: Checkout
+          uses: actions/checkout@v3
+
+        - name: Install dependencies
+          run: |
+            python -m pip install --upgrade pip
+            python -m pip install pytest
+            pip install .
+
+        - name: Test with pytest
+          run: |
+            python -m pytest test/test_process_hac.py
+  
+  CAFE:
+    runs-on: ubuntu-latest
+
+      container:
+        image: docker://devkitpro/devkitppc
+        options: --cpus 2
+
+      steps:
+        - name: Checkout
+          uses: actions/checkout@v3
+
+        - name: Install dependencies
+          run: |
+            python -m pip install --upgrade pip
+            python -m pip install pytest
+            pip install .
+        
+        - name: Test with pytest
+          run: |
+            python -m pytest test/test_process_ctr.py
